Moved reader class to the python scripts and fixed the imports
