+++ applications/GeoMechanicsApplication/tests/c_phi_reduction_process.py
@@ -2,6 +2,7 @@ import os
 import KratosMultiphysics as Kratos
 import KratosMultiphysics.KratosUnittest as KratosUnittest
 from KratosMultiphysics.GeoMechanicsApplication.geomechanics_analysis import GeoMechanicsAnalysis
+from KratosMultiphysics.GeoMechanicsApplication.gid_output_file_reader import GiDOutputFileReader
 import KratosMultiphysics.GeoMechanicsApplication.run_multiple_stages as run_multiple_stages
 
 import test_helper
@@ -17,7 +18,7 @@ class KratosGeoMechanicsCPhiReductionProcess(KratosUnittest.TestCase):
         run_multiple_stages.run_stages(file_path, 2)
 
         # read results
-        reader = test_helper.GiDOutputFileReader()
+        reader = GiDOutputFileReader()
         actual_data = reader.read_output_from(os.path.join(file_path, "stage2.post.res"))
         node_ids = [57, 151, 235]
         displacement_vectors = reader.nodal_values_at_time("DISPLACEMENT", 0.2, actual_data, node_ids)
