+++ applications/GeoMechanicsApplication/tests/test_dsettlement_validation.py
@@ -1,3 +1,4 @@
+from KratosMultiphysics.GeoMechanicsApplication.gid_output_file_reader import GiDOutputFileReader
 from KratosMultiphysics.GeoMechanicsApplication import run_multiple_stages
 from KratosMultiphysics.GeoMechanicsApplication import unit_conversions
 import KratosMultiphysics.KratosUnittest as KratosUnittest
@@ -33,7 +34,7 @@ def get_nodal_vertical_effective_stress_at_time(time_in_s, output_data, node_ids
     return make_compression_positive_and_convert_Pa_to_kPa(
         [
             stress_vector[1]
-            for stress_vector in test_helper.GiDOutputFileReader.nodal_values_at_time(
+            for stress_vector in GiDOutputFileReader.nodal_values_at_time(
                 "CAUCHY_STRESS_TENSOR", time_in_s, output_data, node_ids=node_ids
             )
         ]
@@ -44,7 +45,7 @@ def get_nodal_vertical_total_stress_at_time(time_in_s, output_data, node_ids=Non
     return make_compression_positive_and_convert_Pa_to_kPa(
         [
             stress_vector[1]
-            for stress_vector in test_helper.GiDOutputFileReader.nodal_values_at_time(
+            for stress_vector in GiDOutputFileReader.nodal_values_at_time(
                 "TOTAL_STRESS_TENSOR", time_in_s, output_data, node_ids=node_ids
             )
         ]
@@ -53,7 +54,7 @@ def get_nodal_vertical_total_stress_at_time(time_in_s, output_data, node_ids=Non
 
 def get_nodal_water_pressures_at_time(time_in_s, output_data, node_ids=None):
     return make_compression_positive_and_convert_Pa_to_kPa(
-        test_helper.GiDOutputFileReader.nodal_values_at_time(
+        GiDOutputFileReader.nodal_values_at_time(
             "WATER_PRESSURE", time_in_s, output_data, node_ids=node_ids
         )
     )
@@ -211,7 +212,7 @@ class KratosGeoMechanicsDSettlementValidationTests(KratosUnittest.TestCase):
         project_path = pathlib.Path(project_path)
         ref_path = pathlib.Path(ref_path)
 
-        reader = test_helper.GiDOutputFileReader()
+        reader = GiDOutputFileReader()
 
         output_stage_3 = reader.read_output_from(project_path / "stage3.post.res")
         output_stage_5 = reader.read_output_from(project_path / "stage5.post.res")
@@ -363,7 +364,7 @@ class KratosGeoMechanicsDSettlementValidationTests(KratosUnittest.TestCase):
         project_path = pathlib.Path(project_path)
         ref_path = pathlib.Path(ref_path)
 
-        reader = test_helper.GiDOutputFileReader()
+        reader = GiDOutputFileReader()
 
         output_stage_3 = reader.read_output_from(project_path / "stage3.post.res")
         output_stage_5 = reader.read_output_from(project_path / "stage5.post.res")
@@ -515,7 +516,7 @@ class KratosGeoMechanicsDSettlementValidationTests(KratosUnittest.TestCase):
         )
         self.assertEqual(status, 0)
 
-        reader = test_helper.GiDOutputFileReader()
+        reader = GiDOutputFileReader()
 
         output_data = reader.read_output_from(
             os.path.join(project_path, "stage3.post.res")
@@ -565,7 +566,7 @@ class KratosGeoMechanicsDSettlementValidationTests(KratosUnittest.TestCase):
         )
         self.assertEqual(status, 0)
 
-        reader = test_helper.GiDOutputFileReader()
+        reader = GiDOutputFileReader()
 
         output_data = reader.read_output_from(
             os.path.join(project_path, "stage3.post.res")
@@ -617,7 +618,7 @@ class KratosGeoMechanicsDSettlementValidationTests(KratosUnittest.TestCase):
 
         project_path = pathlib.Path(project_path)
 
-        reader = test_helper.GiDOutputFileReader()
+        reader = GiDOutputFileReader()
 
         output_stage_2 = reader.read_output_from(project_path / "stage2.post.res")
         output_stage_5 = reader.read_output_from(project_path / "stage5.post.res")
@@ -776,7 +777,7 @@ class KratosGeoMechanicsDSettlementValidationTests(KratosUnittest.TestCase):
         )
         self.assertEqual(status, 0)
 
-        reader = test_helper.GiDOutputFileReader()
+        reader = GiDOutputFileReader()
         top_node_ids = [2, 3, 104]
         project_path = pathlib.Path(project_path)
 
@@ -993,7 +994,7 @@ class KratosGeoMechanicsDSettlementValidationTests(KratosUnittest.TestCase):
 
         project_path = pathlib.Path(project_path)
 
-        reader = test_helper.GiDOutputFileReader()
+        reader = GiDOutputFileReader()
 
         output_stage_2 = reader.read_output_from(project_path / "stage2.post.res")
         output_stage_5 = reader.read_output_from(project_path / "stage5.post.res")
