+++ applications/GeoMechanicsApplication/tests/test_elements.py
@@ -2,6 +2,7 @@ import sys
 import os
 import json
 import KratosMultiphysics.KratosUnittest as KratosUnittest
+from KratosMultiphysics.GeoMechanicsApplication.gid_output_file_reader import GiDOutputFileReader
 import test_helper
 
 class KratosGeoMechanicsElementTypeTests(KratosUnittest.TestCase):
@@ -195,7 +196,7 @@ class KratosGeoMechanicsElementTypeTests(KratosUnittest.TestCase):
             effective_stresses_zz = [integration_point[2,2] for element in effective_stresses for integration_point in element]
 
         end_time = 1.0
-        displacements = test_helper.GiDOutputFileReader.nodal_values_at_time("DISPLACEMENT", end_time, output_data)
+        displacements = GiDOutputFileReader.nodal_values_at_time("DISPLACEMENT", end_time, output_data)
         x_displacements = [displacement[0] for displacement in displacements]
         if n_dim >= 2:
             y_displacements = [displacement[1] for displacement in displacements]
@@ -243,12 +244,12 @@ class KratosGeoMechanicsElementTypeTests(KratosUnittest.TestCase):
                 self.assertAlmostEqual(0.0, z_displacement)
 
     def fetchOutputFromFile(self, output_file_path):
-        reader = test_helper.GiDOutputFileReader()
+        reader = GiDOutputFileReader()
         return reader.read_output_from(output_file_path)
 
     def assertVerticalStressAtBottomNodes(self, output_data, bottom_node_ids):
         end_time = 1.0
-        nodal_stress_tensors = test_helper.GiDOutputFileReader.nodal_values_at_time("CAUCHY_STRESS_TENSOR", end_time, output_data, bottom_node_ids)
+        nodal_stress_tensors = GiDOutputFileReader.nodal_values_at_time("CAUCHY_STRESS_TENSOR", end_time, output_data, bottom_node_ids)
         expected_stress_yy = -1e4
         self.assertEqual(len(bottom_node_ids), len(nodal_stress_tensors))
         for stress_tensor in nodal_stress_tensors:
