+++ applications/GeoMechanicsApplication/tests/test_integration_node_extrapolation.py
@@ -1,6 +1,7 @@
 import os
 
 import KratosMultiphysics.KratosUnittest as KratosUnittest
+from KratosMultiphysics.GeoMechanicsApplication.gid_output_file_reader import GiDOutputFileReader
 import test_helper
 
 
@@ -22,9 +23,9 @@ class KratosGeoMechanicsExtrapolationTests(KratosUnittest.TestCase):
         test_name  = '3_noded_triangle'
         file_path  = test_helper.get_file_path(os.path.join(directory, test_name))
         test_helper.run_kratos(file_path)
-        reader     = test_helper.GiDOutputFileReader()
+        reader     = GiDOutputFileReader()
         simulation_output = reader.read_output_from(os.path.join(file_path, test_name+'.post.res'))
-        heads             = test_helper.GiDOutputFileReader.nodal_values_at_time("HYDRAULIC_HEAD", 1, simulation_output,
+        heads             = GiDOutputFileReader.nodal_values_at_time("HYDRAULIC_HEAD", 1, simulation_output,
                                                                              node_ids=[1, 3, 5])
         expected_heads    = [ 1., 0.5, 0. ]
         for head, expected_head in zip(heads, expected_heads):
@@ -35,9 +36,9 @@ class KratosGeoMechanicsExtrapolationTests(KratosUnittest.TestCase):
         test_name = '4_noded_quadrilateral'
         file_path = test_helper.get_file_path(os.path.join(directory, test_name))
         test_helper.run_kratos(file_path)
-        reader     = test_helper.GiDOutputFileReader()
+        reader     = GiDOutputFileReader()
         simulation_output = reader.read_output_from(os.path.join(file_path, test_name+'.post.res'))
-        heads             = test_helper.GiDOutputFileReader.nodal_values_at_time("HYDRAULIC_HEAD", 1, simulation_output,
+        heads             = GiDOutputFileReader.nodal_values_at_time("HYDRAULIC_HEAD", 1, simulation_output,
                                                                              node_ids=[2, 4])
         expected_heads    = [ 0., 1. ]
         for head, expected_head in zip(heads, expected_heads):
@@ -48,9 +49,9 @@ class KratosGeoMechanicsExtrapolationTests(KratosUnittest.TestCase):
         test_name = '6_noded_triangle'
         file_path = test_helper.get_file_path(os.path.join(directory, test_name))
         test_helper.run_kratos(file_path)
-        reader     = test_helper.GiDOutputFileReader()
+        reader     = GiDOutputFileReader()
         simulation_output = reader.read_output_from(os.path.join(file_path, test_name+'.post.res'))
-        heads             = test_helper.GiDOutputFileReader.nodal_values_at_time("HYDRAULIC_HEAD", 1, simulation_output,
+        heads             = GiDOutputFileReader.nodal_values_at_time("HYDRAULIC_HEAD", 1, simulation_output,
                                                                              node_ids=[1, 3, 5])
         expected_heads    = [ 1., 0.5, 0. ]
         for head, expected_head in zip(heads, expected_heads):
@@ -61,16 +62,16 @@ class KratosGeoMechanicsExtrapolationTests(KratosUnittest.TestCase):
         test_name = '8_noded_quadrilateral'
         file_path = test_helper.get_file_path(os.path.join(directory, test_name))
         test_helper.run_kratos(file_path)
-        reader     = test_helper.GiDOutputFileReader()
+        reader     = GiDOutputFileReader()
         simulation_output = reader.read_output_from(os.path.join(file_path, test_name+'.post.res'))
-        heads             = test_helper.GiDOutputFileReader.nodal_values_at_time("HYDRAULIC_HEAD", 1, simulation_output,
+        heads             = GiDOutputFileReader.nodal_values_at_time("HYDRAULIC_HEAD", 1, simulation_output,
                                                                              node_ids=[2, 4, 8])
         expected_heads    = [ 1.09636, 0.292984, 1.29298]
         for head, expected_head in zip(heads, expected_heads):
             self.assertAlmostEqual(head, expected_head, places=4)
 
 
-        cauchy_stress_tensors             = test_helper.GiDOutputFileReader.nodal_values_at_time("CAUCHY_STRESS_TENSOR", 1, simulation_output,
+        cauchy_stress_tensors             = GiDOutputFileReader.nodal_values_at_time("CAUCHY_STRESS_TENSOR", 1, simulation_output,
                                                                                      node_ids=[2, 4, 8])
         expected_tensors    = [ [-1501.6, -6006.41, -1501.6, -376.767, 0, 0],
                                 [733.933, 2935.73, 733.933, -370.876, 0, 0],
@@ -80,7 +81,7 @@ class KratosGeoMechanicsExtrapolationTests(KratosUnittest.TestCase):
                 self.assertAlmostEqual(cauchy_stress, expected, places=None, delta=0.01)
 
 
-        fluid_flux_vectors             = test_helper.GiDOutputFileReader.nodal_values_at_time("FLUID_FLUX_VECTOR", 1, simulation_output,
+        fluid_flux_vectors             = GiDOutputFileReader.nodal_values_at_time("FLUID_FLUX_VECTOR", 1, simulation_output,
                                                                                                          node_ids=[2, 4, 8])
         expected_fluid_flux_vectors    = [ [1.38449e-05, -4.52218e-07, 0],
                             [-3.51008e-06, 2.17395e-05, 0],
