+++ applications/GeoMechanicsApplication/tests/test_prescribed_derivatives.py
@@ -1,6 +1,7 @@
 import os
 
 import KratosMultiphysics.KratosUnittest as KratosUnittest
+from KratosMultiphysics.GeoMechanicsApplication.gid_output_file_reader import GiDOutputFileReader
 
 import test_helper
 
@@ -28,13 +29,13 @@ class KratosGeoMechanicsPrescribedDerivatives(KratosUnittest.TestCase):
         test_helper.run_kratos(file_path)
 
         output_file_path = os.path.join(file_path, test_name + '.post.res')
-        output_reader = test_helper.GiDOutputFileReader()
+        output_reader = GiDOutputFileReader()
         output_data = output_reader.read_output_from(output_file_path)
 
         times = [0.05, 0.1, 0.15, 0.2]
         expected_x_accelerations = [-0.01, -0.005, -0.003, 0.01]
         for expected_x_acceleration, time in zip(expected_x_accelerations, times):
-            accelerations = test_helper.GiDOutputFileReader.nodal_values_at_time("ACCELERATION", time, output_data,
+            accelerations = GiDOutputFileReader.nodal_values_at_time("ACCELERATION", time, output_data,
                                                                                  node_ids=[1, 2, 3])
             self.assertEqual(len(accelerations), 3)
 
@@ -53,13 +54,13 @@ class KratosGeoMechanicsPrescribedDerivatives(KratosUnittest.TestCase):
         test_helper.run_kratos(file_path)
 
         output_file_path = os.path.join(file_path, test_name + '.post.res')
-        output_reader = test_helper.GiDOutputFileReader()
+        output_reader = GiDOutputFileReader()
         output_data = output_reader.read_output_from(output_file_path)
 
         times = [0.05, 0.1, 0.15, 0.2]
         expected_velocities = [-0.01, -0.005, -0.003, 0.01]
         for expected_velocity, time in zip(expected_velocities, times):
-            velocities = test_helper.GiDOutputFileReader.nodal_values_at_time("VELOCITY", time, output_data,
+            velocities = GiDOutputFileReader.nodal_values_at_time("VELOCITY", time, output_data,
                                                                               node_ids=[1, 2, 3])
 
             self.assertEqual(len(velocities), 3)
@@ -78,13 +79,13 @@ class KratosGeoMechanicsPrescribedDerivatives(KratosUnittest.TestCase):
         test_helper.run_kratos(file_path)
 
         output_file_path = os.path.join(file_path, test_name + '.post.res')
-        output_reader = test_helper.GiDOutputFileReader()
+        output_reader = GiDOutputFileReader()
         output_data = output_reader.read_output_from(output_file_path)
 
         times = [0.05, 0.1, 0.15, 0.2]
         expected_dt_water_pressures = [-0.01, -0.005, -0.003, 0.01]
         for expected_dt_water_pressure, time in zip(expected_dt_water_pressures, times):
-            dt_water_pressures = test_helper.GiDOutputFileReader.nodal_values_at_time("DT_WATER_PRESSURE", time,
+            dt_water_pressures = GiDOutputFileReader.nodal_values_at_time("DT_WATER_PRESSURE", time,
                                                                                       output_data,
                                                                                       node_ids=[1, 2, 3])
             self.assertEqual(len(dt_water_pressures), 3)
