+++ applications/GeoMechanicsApplication/tests/test_truss_backbone_mat.py
@@ -1,6 +1,7 @@
 import os
 
 import KratosMultiphysics.KratosUnittest as KratosUnittest
+from KratosMultiphysics.GeoMechanicsApplication.gid_output_file_reader import GiDOutputFileReader
 import test_helper
 
 
@@ -30,7 +31,7 @@ class KratosGeoMechanicsTrussBackboneMaterialTests(KratosUnittest.TestCase):
 
         # output
         output_file_name = os.path.join(project_path, 'truss_backbone_mat.post.res')
-        reader = test_helper.GiDOutputFileReader()
+        reader = GiDOutputFileReader()
         output_data = reader.read_output_from(output_file_name)
 
         times = [1.0, 2.0, 3.0, 4.0]
@@ -38,9 +39,9 @@ class KratosGeoMechanicsTrussBackboneMaterialTests(KratosUnittest.TestCase):
         expected_displacements_x = [1.0, 0.75, 1.0, 2.0]
         for time, expected_force_x, expected_displacement_x in zip(times, expected_forces_x, expected_displacements_x):
             # integration point check in element 1, integration point 1 ( uniform stress and strain so an arbitrary choice )
-            section_force = test_helper.GiDOutputFileReader.element_integration_point_values_at_time("FORCE", time, output_data, [1], [0])[0][0]
+            section_force = GiDOutputFileReader.element_integration_point_values_at_time("FORCE", time, output_data, [1], [0])[0][0]
             self.assertAlmostEqual(section_force[0], expected_force_x, 2)
-            displacement = test_helper.GiDOutputFileReader.nodal_values_at_time("DISPLACEMENT", time, output_data, [3])[0]
+            displacement = GiDOutputFileReader.nodal_values_at_time("DISPLACEMENT", time, output_data, [3])[0]
             self.assertAlmostEqual(displacement[0], expected_displacement_x, 2)
 
     def test_truss_backbone_mat_compression(self):
@@ -56,7 +57,7 @@ class KratosGeoMechanicsTrussBackboneMaterialTests(KratosUnittest.TestCase):
 
         # output
         output_file_name = os.path.join(project_path, 'truss_backbone_mat.post.res')
-        reader = test_helper.GiDOutputFileReader()
+        reader = GiDOutputFileReader()
         output_data = reader.read_output_from(output_file_name)
 
         times = [1.0, 2.0, 3.0, 4.0]
@@ -64,9 +65,9 @@ class KratosGeoMechanicsTrussBackboneMaterialTests(KratosUnittest.TestCase):
         expected_displacements_x = [-1.0, -0.75, -1.0, -2.0]
         for time, expected_force_x, expected_displacement_x in zip(times, expected_forces_x, expected_displacements_x):
             # integration point check in element 1, integration point 1 ( uniform stress and strain so an arbitrary choice )
-            section_force = test_helper.GiDOutputFileReader.element_integration_point_values_at_time("FORCE", time, output_data, [1], [0])[0][0]
+            section_force = GiDOutputFileReader.element_integration_point_values_at_time("FORCE", time, output_data, [1], [0])[0][0]
             self.assertAlmostEqual(section_force[0], expected_force_x, 2)
-            displacement = test_helper.GiDOutputFileReader.nodal_values_at_time("DISPLACEMENT", time, output_data, [3])[0]
+            displacement = GiDOutputFileReader.nodal_values_at_time("DISPLACEMENT", time, output_data, [3])[0]
             self.assertAlmostEqual(displacement[0], expected_displacement_x, 2)
 
     def test_truss_backbone_mat_tension_compression(self):
@@ -82,7 +83,7 @@ class KratosGeoMechanicsTrussBackboneMaterialTests(KratosUnittest.TestCase):
 
         # output
         output_file_name = os.path.join(project_path, 'truss_backbone_mat.post.res')
-        reader = test_helper.GiDOutputFileReader()
+        reader = GiDOutputFileReader()
         output_data = reader.read_output_from(output_file_name)
 
         times = [1.0, 2.0, 3.0, 4.0, 5.0]
@@ -90,9 +91,9 @@ class KratosGeoMechanicsTrussBackboneMaterialTests(KratosUnittest.TestCase):
         expected_displacements_x = [1.0, 0.0, -1.0, 0.0, 1.0]
         for time, expected_force_x, expected_displacement_x in zip(times, expected_forces_x, expected_displacements_x):
             # integration point check in element 1, integration point 1 ( uniform stress and strain so an arbitrary choice )
-            section_force = test_helper.GiDOutputFileReader.element_integration_point_values_at_time("FORCE", time, output_data, [1], [0])[0][0]
+            section_force = GiDOutputFileReader.element_integration_point_values_at_time("FORCE", time, output_data, [1], [0])[0][0]
             self.assertAlmostEqual(section_force[0], expected_force_x, 2)
-            displacement = test_helper.GiDOutputFileReader.nodal_values_at_time("DISPLACEMENT", time, output_data, [3])[0]
+            displacement = GiDOutputFileReader.nodal_values_at_time("DISPLACEMENT", time, output_data, [3])[0]
             self.assertAlmostEqual(displacement[0], expected_displacement_x, 2)
 
     def test_truss_backbone_mat_compression_tension(self):
@@ -108,7 +109,7 @@ class KratosGeoMechanicsTrussBackboneMaterialTests(KratosUnittest.TestCase):
 
         # output
         output_file_name = os.path.join(project_path, 'truss_backbone_mat.post.res')
-        reader = test_helper.GiDOutputFileReader()
+        reader = GiDOutputFileReader()
         output_data = reader.read_output_from(output_file_name)
 
         times = [1.0, 2.0, 3.0, 4.0, 5.0]
@@ -116,9 +117,9 @@ class KratosGeoMechanicsTrussBackboneMaterialTests(KratosUnittest.TestCase):
         expected_displacements_x = [-1.0, 0.0, 1.0, 0.0, -1.0]
         for time, expected_force_x, expected_displacement_x in zip(times, expected_forces_x, expected_displacements_x):
             # integration point check in element 1, integration point 1 ( uniform stress and strain so an arbitrary choice )
-            section_force = test_helper.GiDOutputFileReader.element_integration_point_values_at_time("FORCE", time, output_data, [1], [0])[0][0]
+            section_force = GiDOutputFileReader.element_integration_point_values_at_time("FORCE", time, output_data, [1], [0])[0][0]
             self.assertAlmostEqual(section_force[0], expected_force_x, 2)
-            displacement = test_helper.GiDOutputFileReader.nodal_values_at_time("DISPLACEMENT", time, output_data, [3])[0]
+            displacement = GiDOutputFileReader.nodal_values_at_time("DISPLACEMENT", time, output_data, [3])[0]
             self.assertAlmostEqual(displacement[0], expected_displacement_x, 2)
 
 if __name__ == '__main__':
